pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-wealth-manager"
version = "1.0.0"
description = "AI/ML-Driven Wealth Management System for Stock Market Prediction"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "AI Wealth Team"}
]
keywords = [
    "stock-market",
    "prediction",
    "machine-learning",
    "deep-learning",
    "wealth-management",
    "indian-stocks",
    "nse",
    "bse"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "numpy>=2.0.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "tensorflow>=2.15.0",
    "keras>=3.0.0",
    "torch>=2.0.0",
    "yfinance>=0.2.0",
    "ta>=0.10.0",
    "pandas-ta>=0.3.14b0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.15.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "tabulate>=0.9.0",
    "xgboost>=2.0.0",
    "lightgbm>=4.0.0",
    "optuna>=3.3.0",
    "prophet>=1.1.0",
    "statsmodels>=0.14.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "joblib>=1.3.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.65.0",
    "scipy>=1.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "isort>=5.12.0",
]

[project.scripts]
wealth = "src.cli.main:cli"

[project.urls]
Homepage = "https://github.com/Joy-Macwan/stock_market_prediction"
Repository = "https://github.com/Joy-Macwan/stock_market_prediction"
Documentation = "https://github.com/Joy-Macwan/stock_market_prediction#readme"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing"
